refactor(apple): Use `kSecUseDataProtectionKeychain` for token (#7756)
This will force the macOS Keychain to behave like the iOS Keychain. To
be honest, Apple's documentation is very much lacking in this regard,
but some research suggests this is both heavily recommended by Apple and
that it enables the Keychain operation to benefit from Apple's security
hardware in their Macs.

In my local testing, it also seems to make keychain operations more
reliable when SIP is disabled, but that could be a fluke given the
number of variables at play.


https://developer.apple.com/documentation/security/ksecusedataprotectionkeychain

Draft because stacked.