fix(connlib): don't drop unsent datagrams (#7768)
We introduced a regression in `connlib` in #7749 whereby queued but
unsent datagrams got dropped in case the socket was not ready to send
more data.

This happens because within `Io`, we pull each datagram one by one from
the iterator:
https://github.com/firezone/firezone/blob/e60ec7144cfb38c9fbbc26cd3d28e1f41665dc58/rust/connlib/tunnel/src/io.rs#L178-L188

This function will send datagrams for as long as the socket is ready and
drop the iterator afterwards. This means the returned iterator MUST BE
lazy and "cancel-safe". This was the case prior to #7749 because
`datagrams` function used `iter_mut` and only cut off the to be sent
bytes when the next item got pulled from iterator. With #7749, the
entire `HashMap` got drained, thus dropping packets if `Io` didn't
manage to process the iterator in full.