ci: Only upload Tauri builds on workflow_dispatch (#7795)
Similar to the Apple and Android clients, this PR updates the Linux and
Windows GUI clients to upload to the GitHub drafted release on manual
workflow triggers only.

This should save a few minutes off `main` builds as the extra package
testing steps will now be skipped there.

Notably, the Gateway and Headless Client workflows are unchanged because
(a) they are much faster to build / test and (b) we use the release
builds for performance testing connlib, so we need them to run on
`main`.