fix(apple/macOS): Memoize successful SCDynamicStoreCreate (#7892)
Under some conditions the call to SCDynamicStoreCreate can fail,
presumably due to some kind of allocation failure. Once it succeeds,
however, we can continue using the dynamic store for the lifetime of the
Adapter instance.

So we memoize the result of this call so as not to allocate each time we
need it.

See
https://developer.apple.com/documentation/systemconfiguration/1437828-scdynamicstorecreate