ci: don't reference old Sentry project name (#7960)
We have recently merged the two Sentry projects for the GUI client into
just one, i.e. both the GUI client and its IPC service use the same DSN
to report errors.

What we forgot to change is that these are also referenced in the CI
workflow when we upload debug symbols.