refactor(gui-client): simplify IPC message handling code (#7955)
At present, the GUI client uses a separate task for reading messages
from the IPC connection and forwards them to another channel. The other
end of this channel is then used within the controller to actually react
to IPC messages.

We can simplify this by removing the intermediary task and processing
the messages from the IPC connection directly.