chore(ops): Update CPU utilization alert threshold (#7966)
The PR updates the threshold for CPU utilization monitoring alerts. This
is being done to avoid notification fatigue. This is not intended to
ignore any issue that might cause the CPU utilization to be high, but
rather make sure we don't miss other important alerts that might come
through while we try to solve the underlying CPU utilization issues.