fix(gui-client): don't fail DNS control on nameservers (#7972)
As part of fixing #6777, we added a code path to explicitly set
nameservers on the tunnel interface. This was necessary to make DNS
resources work under WSL. Exactly this code also seems to be causing
issues where this particular registry key is not available on a users
machine.

DNS resources outside of WSL will also work without this, therefore we
make this part optional to at least have something working on the users
machine.

Related: #7962.