chore(portal): Remove GCP alerting for application errors (#8040)
As discussed with @bmanifold, we're moving forward with the following
monitoring strategy:

For infra alerts, stick with GCP.

For application-level alerts, use Sentry.

Since we already have Sentry configured and working on staging, this PR
removes the "Errors in logs" alert since we will be receiving this in
Sentry going forward.